+++ 05-travel-list/starter/src/components/Logo.js
@@ -1,0 +1,3 @@
+export default function Logo() {
+  return <h1>🌴 Far Away 💼</h1>;
+}
